refactor: modularize TypeScript files over threshold
