+++ frontends/nextjs/src/types/dbal/blob.d.ts
@@ -1,0 +1,36 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+declare module '@/dbal/development/src/blob' {
+  export interface BlobStorageConfig {
+    type: 'filesystem' | 'memory' | 's3'
+    basePath?: string
+  }
+
+  export interface BlobMetadata {
+    contentType?: string
+    size?: number
+    lastModified?: Date
+    [key: string]: any
+  }
+
+  export interface BlobListItem {
+    key: string
+    [key: string]: any
+  }
+
+  export interface BlobListResult {
+    items: BlobListItem[]
+    [key: string]: any
+  }
+
+  export interface BlobStorage {
+    upload(key: string, data: Buffer | string, metadata?: BlobMetadata): Promise<string>
+    download(key: string): Promise<Buffer>
+    delete(key: string): Promise<void>
+    exists(key: string): Promise<boolean>
+    list(options?: { prefix?: string }): Promise<BlobListResult>
+    getMetadata(key: string): Promise<BlobMetadata | null>
+  }
+
+  export function createBlobStorage(config: BlobStorageConfig): BlobStorage
+}
