+++ frontends/nextjs/src/types/dbal/core-types.d.ts
@@ -1,0 +1,11 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+declare module '@/dbal/development/src/core/types' {
+  export interface User {
+    id: string
+    email: string
+    name?: string
+    level: number
+    tenantId: string
+  }
+}
